requirements.txt
flask>=3.1.0
python-dotenv>=1.0.1
firebase-admin>=6.6.0
openpyxl>=3.1.5